+++ src/sentry/auth/services/orgauthtoken/impl.py
@@ -3,8 +3,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from sentry.auth.services.orgauthtoken.service import OrgAuthTokenService
+from sentry.hybridcloud.outbox.category import OutboxCategory, OutboxScope
 from sentry.models.orgauthtoken import OrgAuthToken
-from sentry.models.outbox import OutboxCategory, OutboxScope, RegionOutbox, outbox_context
+from sentry.models.outbox import RegionOutbox, outbox_context
 
 if TYPE_CHECKING:
     from datetime import datetime
