+++ src/sentry/integrations/models/external_actor.py
@@ -9,9 +9,9 @@ from sentry.backup.scopes import RelocationScope
 from sentry.db.models import BoundedPositiveIntegerField, FlexibleForeignKey, region_silo_model
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
 from sentry.hybridcloud.outbox.base import ReplicatedRegionModel
+from sentry.hybridcloud.outbox.category import OutboxCategory
 from sentry.hybridcloud.services.replica import control_replica_service
 from sentry.integrations.types import ExternalProviders
-from sentry.models.outbox import OutboxCategory
 from sentry.notifications.services import notifications_service
 
 logger = logging.getLogger(__name__)
