+++ src/sentry/models/apikey.py
@@ -12,9 +12,9 @@ from sentry.db.models import BoundedPositiveIntegerField, control_silo_model, sa
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
 from sentry.db.models.manager.base import BaseManager
 from sentry.hybridcloud.outbox.base import ReplicatedControlModel
+from sentry.hybridcloud.outbox.category import OutboxCategory
 from sentry.hybridcloud.services.replica import region_replica_service
 from sentry.models.apiscopes import HasApiScopes
-from sentry.models.outbox import OutboxCategory
 
 
 # TODO(dcramer): pull in enum library
