+++ src/sentry/models/apitoken.py
@@ -18,9 +18,9 @@ from sentry.backup.scopes import ImportScope, RelocationScope
 from sentry.constants import SentryAppStatus
 from sentry.db.models import FlexibleForeignKey, control_silo_model, sane_repr
 from sentry.hybridcloud.outbox.base import ControlOutboxProducingManager, ReplicatedControlModel
+from sentry.hybridcloud.outbox.category import OutboxCategory
 from sentry.models.apigrant import ApiGrant
 from sentry.models.apiscopes import HasApiScopes
-from sentry.models.outbox import OutboxCategory
 from sentry.types.region import find_all_region_names
 from sentry.types.token import AuthTokenType
