+++ src/sentry/models/authprovider.py
@@ -21,7 +21,8 @@ from sentry.db.models import (
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
 from sentry.db.models.fields.jsonfield import JSONField
 from sentry.hybridcloud.outbox.base import ReplicatedControlModel
-from sentry.models.outbox import ControlOutbox, OutboxCategory, OutboxScope
+from sentry.hybridcloud.outbox.category import OutboxCategory, OutboxScope
+from sentry.models.outbox import ControlOutbox
 from sentry.types.region import find_regions_for_orgs
 
 logger = logging.getLogger("sentry.authprovider")
