+++ src/sentry/models/organizationmemberteam.py
@@ -9,7 +9,8 @@ from sentry import features, roles
 from sentry.backup.scopes import RelocationScope
 from sentry.db.models import BoundedAutoField, FlexibleForeignKey, region_silo_model, sane_repr
 from sentry.hybridcloud.outbox.base import RegionOutboxProducingManager, ReplicatedRegionModel
-from sentry.models.outbox import OutboxCategory, RegionOutboxBase
+from sentry.hybridcloud.outbox.category import OutboxCategory
+from sentry.models.outbox import RegionOutboxBase
 from sentry.roles import team_roles
 from sentry.roles.manager import TeamRole
