+++ src/sentry/models/useremail.py
@@ -22,7 +22,8 @@ from sentry.backup.scopes import ImportScope, RelocationScope
 from sentry.db.models import FlexibleForeignKey, control_silo_model, sane_repr
 from sentry.db.models.manager.base import BaseManager
 from sentry.hybridcloud.outbox.base import ControlOutboxProducingModel
-from sentry.models.outbox import ControlOutboxBase, OutboxCategory
+from sentry.hybridcloud.outbox.category import OutboxCategory
+from sentry.models.outbox import ControlOutboxBase
 from sentry.organizations.services.organization.model import RpcOrganization
 from sentry.types.region import find_regions_for_user
 from sentry.users.services.user.model import RpcUser
