+++ src/sentry/models/userrole.py
@@ -11,7 +11,8 @@ from sentry.backup.scopes import RelocationScope
 from sentry.db.models import ArrayField, control_silo_model, sane_repr
 from sentry.db.models.fields.foreignkey import FlexibleForeignKey
 from sentry.hybridcloud.outbox.base import ControlOutboxProducingModel
-from sentry.models.outbox import ControlOutboxBase, OutboxCategory
+from sentry.hybridcloud.outbox.category import OutboxCategory
+from sentry.models.outbox import ControlOutboxBase
 from sentry.signals import post_upgrade
 from sentry.silo.base import SiloMode
 from sentry.types.region import find_all_region_names
