+++ src/sentry/relocation/services/relocation_export/impl.py
@@ -11,9 +11,10 @@ from uuid import UUID
 from django.db import router
 from sentry_sdk import capture_exception
 
+from sentry.hybridcloud.outbox.category import OutboxCategory, OutboxScope
 from sentry.models.files.file import File
 from sentry.models.files.utils import get_relocation_storage
-from sentry.models.outbox import ControlOutbox, OutboxCategory, OutboxScope
+from sentry.models.outbox import ControlOutbox
 from sentry.models.relocation import Relocation, RelocationFile
 from sentry.relocation.services.relocation_export.model import (
     RelocationExportReplyWithExportParameters,
