+++ tests/sentry/middleware/integrations/parsers/test_github.py
@@ -4,10 +4,11 @@ from django.http import HttpRequest, HttpResponse
 from django.test import RequestFactory, override_settings
 from django.urls import reverse
 
+from sentry.hybridcloud.outbox.category import OutboxCategory
 from sentry.integrations.models.integration import Integration
 from sentry.integrations.models.organization_integration import OrganizationIntegration
 from sentry.middleware.integrations.parsers.github import GithubRequestParser
-from sentry.models.outbox import ControlOutbox, OutboxCategory, outbox_context
+from sentry.models.outbox import ControlOutbox, outbox_context
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import TestCase
 from sentry.testutils.outbox import assert_no_webhook_payloads, assert_webhook_payloads_for_mailbox
