+++ tests/sentry/middleware/integrations/parsers/test_gitlab.py
@@ -7,11 +7,12 @@ from django.test import RequestFactory, override_settings
 from django.urls import reverse
 
 from fixtures.gitlab import EXTERNAL_ID, PUSH_EVENT, WEBHOOK_SECRET, WEBHOOK_TOKEN
+from sentry.hybridcloud.outbox.category import OutboxCategory
 from sentry.integrations.models.integration import Integration
 from sentry.integrations.models.organization_integration import OrganizationIntegration
 from sentry.middleware.integrations.classifications import IntegrationClassification
 from sentry.middleware.integrations.parsers.gitlab import GitlabRequestParser
-from sentry.models.outbox import ControlOutbox, OutboxCategory, outbox_context
+from sentry.models.outbox import ControlOutbox, outbox_context
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import TestCase
 from sentry.testutils.outbox import assert_no_webhook_payloads, assert_webhook_payloads_for_mailbox
