+++ tests/sentry/models/test_sentryapp.py
@@ -1,7 +1,8 @@
 from sentry.constants import SentryAppStatus
+from sentry.hybridcloud.outbox.category import OutboxCategory
 from sentry.models.apiapplication import ApiApplication
 from sentry.models.integrations.sentry_app import SentryApp
-from sentry.models.outbox import ControlOutbox, OutboxCategory
+from sentry.models.outbox import ControlOutbox
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
